Revert "FilesStorageProvider service (#27)" (#28)
This reverts commit c1b97c0cc99b82e467838f9f84eb450723c6a3bb.
