+++ SB-MDEditor/App/Di/Di+ModuleFactory.swift
@@ -25,3 +25,46 @@ protocol ModuleFactory: AnyObject {
 	func makeOpenDocModule() -> Module
 	func makeCreateDocModule() -> Module
 }
+
+extension Di: ModuleFactory {
+	func makeStartModule() -> Module {
+		makeMainModule()
+	}
+
+	func makeAboutModule() -> Module {
+		let viewController = AboutViewController()
+		let navigationVC = UINavigationController(rootViewController: viewController)
+		navigationVC.navigationBar.prefersLargeTitles = true
+		return .init(viewController: navigationVC, animatedType: .fade)
+	}
+
+	func makeMainModule() -> Module {
+		let presenter = MainPresenter()
+		let worker = MainWorker()
+		let interactor = MainInteractor(presenter: presenter, worker: worker)
+		let router = MainRouter()
+		let viewController = MainViewController(interactor: interactor, router: router)
+
+		presenter.viewController = viewController
+		router.viewController = viewController
+		router.dataStore = interactor
+		router.view = rootVC
+
+		let navigationVC = UINavigationController(rootViewController: viewController)
+		navigationVC.navigationBar.prefersLargeTitles = true
+
+		return .init(viewController: navigationVC)
+	}
+
+	func makeOpenDocModule() -> Module {
+		let viewController = OpenDocViewController()
+		let navigationVC = UINavigationController(rootViewController: viewController)
+		navigationVC.navigationBar.isHidden = true
+		return .init(viewController: navigationVC, animatedType: .dismiss)
+	}
+
+	func makeCreateDocModule() -> Module {
+		let viewController = CreateDocViewController()
+		return .init(viewController: viewController, animatedType: .fade)
+	}
+}
