--- SB-MDEditor/App/Di/Di+ServiceFactory.swift
@@ -1,17 +1,0 @@
-import Foundation
-
-// MARK: - ServicesFactory
-
-protocol ServicesFactory {
-	func makeFilesStorageProvider(_ storage: UserDefaults) -> IFilesStorageProvider
-}
-
-extension Di: ServicesFactory {
-	func makeFilesStorageProvider(_ storage: UserDefaults) -> IFilesStorageProvider {
-		FilesStorageProvider(
-			userDefaults: storage,
-			bandlePath: "",
-			aboutPath: ""
-		)
-	}
-}
